#32 Progress Deleted some files, refactor directories structure, update .gitignore and add uv as dependecies manager
