+++ Src/Logging/logger_config_debug.json
@@ -1,0 +1,3 @@
+{
+    "level": 10
+}
