--- Src/Themes/__init__.py
@@ -1,1 +1,0 @@
-from Src.Themes.theme_manager import ThemeManager
