+++ main.py
@@ -1,25 +1,28 @@
 import json
 import sys
+from pathlib import Path
 
 import dearpygui.dearpygui as dpg
 
 from Src.Logging import Logger_factory
-from Src.Themes import ThemeManager
+from Src.Managers import ThemeManager
 from Src.node_editor import NodeEditor
 
 
 dpg.create_context()
 dpg.create_viewport(title='Custom Title')
 
-base_path = sys._MEIPASS if hasattr(sys, '_MEIPASS') else '.'
-config_path = f"{base_path}/Src/Logging/logger_config.json"
-font_path = f"{base_path}/notomono-regular.ttf"
-ThemeManager.load_themes(f"{base_path}/Src/Themes/themes.json")
+base_path = Path(sys._MEIPASS if hasattr(sys, '_MEIPASS') else '.')
+config_path = base_path / "Src/Logging/logger_config.json"
+font_path = base_path / "Assets/notomono-regular.ttf"
+themes_path = base_path / "Assets/themes.json"
+
 
 with open(config_path) as f:
     config = json.load(f)
 
 log_factory = Logger_factory(config)
+ThemeManager.load_themes(themes_path)
 node_editor = NodeEditor(minimap=True, minimap_location=dpg.mvNodeMiniMap_Location_TopRight)
 main_logger = log_factory("main")
