+++ pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "graphnet"
+version = "0.1.0"
+description = "Add your description here"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "dearpygui>=2.1.0",
+    "numpy>=2.3.4",
+    "pillow>=12.0.0",
+    "tensorflow>=2.20.0",
+]
